make everything `click` commands
